requirements.txt
flask==3.0.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
plotly==5.18.0
anthropic==0.25.0
python-dotenv==1.0.0
reportlab==4.0.7
kaleido==0.2.1
